[ENH]  Safer object store interface. (#3098)
We already wrap up the object store in an interface that doesn't expose
delete.  This allows us to wrap an object store in a non-destructive
interface that advertises when it is non-destructive.  The cache has
been updated to mandate that the cache be destructive and the backing
store be non-destructive.