[ENH] Collection and segment creation in single Tx. (#2928)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
	 - Earlier, a collection and its segments were created as two different transactions. Failure scenarios could result in undesired state. This PR addresses it, and creates collection and segments together in a single transaction.
	 - Changes are made at the implementation of the create_collection api and all layers below it, namely local sysdb, distributed sysdb, grpc implementation of sysdb, and the data layer.
	 - Changes do not include the implementation for delete_collection(). This will be added after this PR merges.

## Test plan
*How are these changes tested?*
Existing tests were modified to reflect the changes. New tests were added as welll.
- [x] Tests pass locally with `pytest` for python, `yarn test` for js, `cargo test` for rust, `make test` for golang.

## Documentation Changes
None


NOTE: This PR contains all the necessary code for the feature, but does not enable the feature. Feature will be enabled in subsequent PR. This is to account for the possibility that a user could use new client to send requests to an older version of server causing issues.