[ENH] Soft-delete Collections (#3070)
## Description of changes

Deleting a collection should not remove the collection, instead just set
the is_deleted field. This will allow manual recovery of collections if needed.
The feature written to be backward compatible - i.e. it should be possible to
turn this feature ON and OFF and the system should not get confused the data
thats written before toggling the feature.
NOTE: Soft Delete is still turned OFF in this PR.

*Summarize the changes made by this PR.*
 - New functionality
- Collections are soft-deleted. i.e. marked as deleted instead of actual
deletion.
- Soft-deleted collections are cleaned up, or actually deleted after a
while.

## Test plan
*How are these changes tested?*
- [ ] Existing Tests pass locally with `pytest` for python, `make test`
for golang.
- [ ] New Tests written to test soft deletion and clean-up of such
collections.

## Documentation Changes
N/A