[BUG] Disallow empty embedding lists in normalization (#3183)
## Description of changes

Caught by @Sicheng-Pan. Prior to this change, it was possible to pass an
empty embeddings list, e.g:

```python
import chromadb
client = chromadb.Client()
coll = client.create_collection("repro")
coll.query(query_embeddings=[])
```

Throws:

```
... in normalize_embeddings
    if isinstance(target[0], (int, float)) and not isinstance(target[0], bool):
                  ~~~~~~^^^
IndexError: list index out of range in query.
```

This is because we normalize before we validate. This is OK for
`query_texts` because the empty list will just get passed to the
validator which will fire correctly, but normalization logic for
embeddings is more complicated so we need to be more defensive.

This change also fixes the same problem for an empty numpy array. 

## Test plan

- CI passes
- No additional tests: tests for validators will be in validators PR. 

## Documentation Changes
N/A