[BLD] Sync internal log service changes to OSS (#3274)
## Description of changes

Currently, we build the log service from a `main.go` that lives in our
private repository. This `main` package imports code from here in
`./go`. Instead of having this weird split world, this syncs the few
differences from the private repository (those from the
`./go/Dockerfile` and using a "sharedOtel" import for the gRPC service)
to here. There is also a change in the private repository to remove all
the dead/duplicated code that we no longer need. This other change also
updates the build to use this code entirely instead of the `main`
package that was defined over there.


## Test plan
- Manual `docker build` and `docker run`
- `tilt up`

- [ ] ~Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust~ not relevant

## Documentation Changes
- not needed