[BUG] Fix HNSW resize conditions to exclude deleted items (#3424)
## Description of changes

We were seeing errors in compaction that looked like:

```
WriteSegments(ApplyMaterializatedLogsError(HnswIndex(FFIException(\"The number of elements exceeds the specified limit\"))))
```
This is because the resize logic uses `len` which removes deleted
elements which then causes the conditional for resizing to be skipped.
This adds a new method `len_with_deleted` to return the length of the
index including the elements that have been deleted so that the resize
is not skipped. See:
https://github.com/chroma-core/chroma/blob/27564afa0e82d184a0fa853a7384d00861d8e9dc/rust/worker/src/segment/distributed_hnsw_segment.rs#L212-L222

## Test plan
A test was added. It will fail when using the existing `len` method, but
passes when using the newly added `len_with_deleted` method.

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
N/A

Thanks to @sanketkedia for helping root cause this. 

Fixes https://github.com/chroma-core/hosted-chroma/issues/855