[BLD] Add link to hotfix PR in action summary (#3433)
## Description of changes

When applying a hotfix, a PR gets generated. This PR is currently only
in the output of the action, or can be found in the list of PRs. This
adds a heading and a link to the PR in the action summary so that the
developer can find it more easily.

## Test plan
I manually linted the action code to make sure I didn't break anything
obvious. I may still want to tweak the output that gets generated, but
at this point, the PR link will be much more easy to spot.