[BUG] Make memberlist use ips for routing (#3405)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- The memberlist uses k8s dns based routing for the stateful set.
However, this can result in delays when the routing updates. This
propagates the ip in the memberlist which is much faster.
   - This change is backwards compatible
- CRDs fields are optional by default, so not setting the ip is fine if
CRD is updated
       - CRD fields can be set in the code without the CRD being updated
- The python frontend will use the ip if present, otherwise it falls
back to the id
- The go code will read the ip to "" anywhere it expects it to be set if
the old version of the CR is used
 - New functionality
   - ...

## Test plan
*How are these changes tested?*
Added a test which adds data to a collection, kills the query service
pods and waits for them to be ready, simulating a roll out of sorts.
Then it issues another query and make sure the query succeeds with
updated routing. Before this change, this test failed.
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None