[BLD] Make artifact names unique for the job being run (#3444)
## Description of changes

We often see failures in this workflow that look like:

```
Error: Failed to CreateArtifact: Received non-retryable error: Failed request: (409) Conflict: an artifact with this name already exists on the workflow run
```

## Test plan
I linted the action, but will need to test it live to see that artifacts
are getting generated appropriately. Expect that many of the failure
conditions we are currently seeing go away.