[ENH] Rust Sqlite migrations (#3599)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
   - /
 - New functionality
   - Adds rust sqlite that copies the python codes migrations system.
- I made the choice to COPY over the migrations from the python code so
that we can add to them if needed only for the rust codebase. We can
backport migrations and be careful to not let these diverge while in
interim development.
- Migration files are embedded with `rust_embed` crate. The design for
the migrations system is not highly generic, but can grow to support
more modules over time, and could theoretically be easily extracted out
for use in a different system with ease.
   - 

## Test plan
*How are these changes tested?*
I added tests for init, the validation and failure cases such as sql,
hash, or version corruption.
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None