[ENH]: suppress prefetch errors when the sender no longer exists (#3664)
## Description of changes

Prefetch operators are "fire and forget"; we don't care if the sender is
no longer available to receive the result of the prefetch.

## Test plan
*How are these changes tested?*

Tested by adding a sleep to a prefetch operator and observing that the
logged event was at the debug level, not error level.

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

n/a