[ENH]  Allow reference work loads.
A reference workload uses two collections.  It references one to get doc
ids and it writes/reads the other with queries, gets, and upserts.

This may sound strange, but it's because we need a canonical way of
knowing what data we're writing and can query so that we can verify
ground truth.

Options considered:
- Load the hugging face data sets to a volume.  Operationally tricky.
- Load the hugging face data sets to a different DB.  Operationally
  tricky.
- Use a reference chroma collection (this).