DOC: fix page overflow issues on mobile
The tabbed-use-case-code-block.tsx would overflow if the width of the
tabs list exceeded the width of the parent. In this case, the copy
button would be outside the tab header. It has been changed to have
the tabs list scroll when overflowing, preventing the sibling elements
from overflowing the header.

In addition, the cloud-signup.tsx was not respecting the viewport width
at narrow widths. This change allows the container to respect the viewport
width up a max width of 400px. Some responsive text sizing was placed on
the copy to prevent text from wrapping and to prevent the text from overflowing
the container in height. Given the way the image onload event is used,
moving it to be a background image is non-trivial and still allow the
same "show once loaded" logic.