[ENH]: use new hnswlib crate (#3782)
## Description of changes

Requires https://github.com/chroma-core/hnswlib/pull/30 (CI will not
pass until that PR is merged).

Moves bindings to our fork of hnswlib. This means that the hnswlib repo
will no longer need to be a sibling of this repo for it to build
correctly.

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

Moved Rust tests for bindings to hnswlib repo.