[BUG]: pin fastapi version (#3872)
## Description of changes

https://github.com/encode/starlette/pull/2812 (transitively updated to
in fastapi 0.115.10) caused large portions of our test suite to fail.
We're pinning fastapi (and thus starlette) for now to the last working
version since we're deprecating the Python server anyways.

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

n/a