[TST]: fix Hypothesis filter generation for floats (#3868)
## Description of changes

Before this change, it was possible for Hypothesis to generate where
filters in which a float was used for `$eq`, `$ne`, `$in`, and `$nin`.
Floats should not be used for these operators as float accuracy errors
can give unexpected results.

Should fix a bad filter generation observed in this test run:
https://github.com/chroma-core/chroma/actions/runs/13578627433/job/37960408763?pr=3868

## Test plan
*How are these changes tested?*

Covered by existing tests.