[ENH] Replace global event sender channel with `ReceiverForMessage<MeterEvent>` (#3898)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- Changed the meter event field names so that the represented value can
be easily aggregated
- Replaced global event sender channel with custom receiver that
implements `ReceiverForMessage<MeterEvent>`. This allows us to spawn a
component in the system and set a global receiver for the meter event
message, and additional logic to listen on the receiver is hidden from
the meter module.
 - New functionality
   - N/A

## Test plan
*How are these changes tested?*

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*