[ENH]: default content-type to application/json if not provided (#3954)
## Description of changes

The Rust HTTP server now defaults the content-type header to
application/json if not provided. This means that Python clients prior
to 1.0.0b are now compatible with the Rust server.

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

Added test was previously failing.

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

n/a