[BUG]: supress not found error on temporary HNSW file removal (#3969)
## Description of changes

In production, we frequently see the error "Failed to remove temporary
files for ******. No such file or directory (os error 2)". Although
unlikely, there's a possible race condition that would cause this: the
index is evicted from the cache while another task is inside the
`purge_by_id()` method.

Changing the not found error to a warning for now to reduce pages.

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

n/a