[ENH] Remove timestamp_us from wal3 LogPosition.
The timestamp_us field was supposed to be helpful, but it leads to
ambiguity when providing offsets to scan.  Chiefly, what should be the
behavior when a LogPosition with a timestamp different from the one
assigned to that offset is used for scan.  This does away with the
ambiguity and reverts to just logging the timestamp of insertion, which
can drift and therefore won't be monotonic.