[ENH]: bump the max response payload size limit for query service clients (#3975)
## Description of changes

Will prevent frontends from erroring on large responses. I added config
parameter for the Rust frontend and hardcoded the limit to 32MB for the
Python frontend since we're removing it soon.

## Test plan
*How are these changes tested?*

Tested by creating 1k documents of 10kB and calling `.get(limit=1000)`.
This threw an error on main and worked on this branch. I can check in
this test if desired, but it's not super straightforward to do in a
clean way since we don't have a good pattern for E2E Rust tests yet.

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

n/a