[ENH]: better axum errors (#4007)
## Description of changes

By default, axum returns some built-in errors (like invalid JSON) as a
plaintext response. This adds middleware so that these errors are
converted to JSON and don't break clients.

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

Added a test that fails on main.

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

n/a