[TST] remove conftest installs and simplify ef tests (#4054)
## Description of changes

there have been consistent issues with embedding function tests. the
root cause was the test fixture used, which attempted to install and
uninstall dependencies required for the tests, and at any point if some
installation did not succeed, it only pushed a log out, before
continuing the tests. therefore tests were guaranteed to fail on any
issue with the installation. this is also slow to have to install &
uninstall packages every time the test runs.

This pr removes the test fixtures and simplifies the mocks used for the
embedding functions