[TST] Fix hnsw degradation for small test cases (#4025)
## Description of changes

*Summarize the changes made by this PR.*
I have confirmed that a persistent source of flakes on main in
test_add/embeddings/persist is due to how we now write directly to HNSW
in the rust code. Under a heavily deleted index, the hsnw graph degrades
and at small scales this means there is no connectivity in the graph.
For example if I have 15 points and delete 13, my graph may have two
distinct connected components of unit size, making search impossible. In
the python code, we used to buffer writes into a brute force layer which
hid this.

My proposal is to under >20% degradation (a heuristic used for rebuild
in other dbs, and <100 scale) to brute force the data.

## Test plan
*How are these changes tested?*
Existing tests catch this case
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None