[ENH] CLI Profile and DB management (#4056)
## Description of changes
* Added DB and profile management commands.
* Introduce `CliError` for better error handling across the CLI. 

### `chroma profile`

Profiles are the way we keep track of logged in states for users. Each
profile contains the API key and tenant ID after authenticating with
Chroma Cloud (with the `login` command).

Available commands:
* `chroma profile use <name>` - sets a profile as the currently active
profile. Errors if the named profile does not exist.
* `chroma profile delete <name>` - deletes a profile. Asks for
confirmation if the profile is the currently active one. Errors if named
profile does not exist.
* `chroma profile list` - lists the available profiles. Informs the user
if no profiles are available.
* `chroma profile show` - displays the currently active profile. Errors
if no active profile is set.
* `chroma profile rename <name>` - rename a profile. Errors if named
profile does not exist.

### `chroma db`

Available commands:
* `chroma db list` - lists the DBs for the logged in profile. Informs
the user if no DBs are available.
* `chroma db create <name>` - create a DB. Errors if name already exists
or is invalid.
* `chroma db delete <name>` - deletes a DB. Errors if name does not
exist.
* `chroma db connect <name>` - generates a code snippet for DB
connection in selected language for the DB. Errors if DB does not exist.

All commands will fail with a relevant message if there is no currently
active profile.

### Testing Locally

1. Build the CLI: `cargo build --bin chroma --release
2. Move the built CLI to `~/.local/bin`
3. Create the following files in `~/.chroma` (handled in the `login`
command, PR coming soon).

```
// credentials

[profile_name]
api_key = "..."
tenant_id = "..."
```

```
// config.json
{
  "current_profile": "profile_name"
}
```

## Test plan
We are using `cmd_assert` to test the CLI. It requires the CLI binary to
built, and allows to easily verify that the correct messages are
displayed to the user.

However, we were not able to stand up a Rust frontend in a test, and
have the built binary used by `cmd_assert` interact with it in the same
test. After discussing with Robert, we decided that it's easier for now
to run such tests as integration tests and debug further later.

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
N/A