[ENH]: write to Foyer disk cache at insert time, fix host path binding for cache volumes in CI (#4079)
## Description of changes

Upgrades foyer to use the new write-to-disk-at-insert-time policy. Also
fixes the host path bindings for the cache volumes so that separate
foyer instances do not write to the same directory.

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

Test asserting that cache could be recovered from disk was previously
failing and is now passing.

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

n/a