[ENH] State machine based property tests for GC. (#4057)
## Description of changes

State machine based property tests for Garbage Collection.
Does not depend on the entire cluster to be up using Tilt.
Uses sparse index based abstractions to add new versions
to the collection.
This is a replacement for the prop tests in 
https://github.com/chroma-core/chroma/pull/3883
That PR will be closed in favor of this.

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
   - Adds property tests that are not dependent on Tilt.
 - New functionality
   - ...

## Test plan
*How are these changes tested?*

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*