[ENH] CLI login command (#4082)
## Description of changes

A CLI command that lets users authenticate with cloud in their browser,
and persist the API key and tenant ID in a "profile" under
`~/.chroma/credentials`.

Expected behavior:
1. `chroma login` (use the `--dev` flag in development environment)
2. Browser opens in trychroma.com/login.
3. After authentication, user gets redirected to a "login success" page.
4. In the terminal, if more than one team exists, the user selects which
team to log into.
5. We create a profile with that team name. If a profile with the same
team name already exists, we ask the user if they want to override it,
or provide a new name.

## Test plan

1. `tilt up` in the hosted repo.
2. `cargo run -- login --dev` in the `rust/cli` directory.
3. Follow the steps above.

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
N/A