[ENH] Include enum -> list of literals for type safety (#4098)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
   - ...
 - New functionality
- Current metadata typing gives a static type checking error in the IDE
when supplied with strings. To fix this, the IncludeEnum type has been
removed and replaced with `Literal[...]` of supported strings. Since
there is no longer official support for python 3.7, we are able to do
this.

## Test plan
*How are these changes tested?*

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*