[ENH]: stop/kill worker threads when dispatcher is stopped (#4102)
## Description of changes

The test `run_gc_test_ext` currently fails on `main` for me when running
locally with the error "too many open files". This is because the worker
threads that our system creates are never cleaned up.

This PR updates `ConsumableJoinHandle` to wrap thread handles, and
updates the dispatcher to stop worker threads when the dispatcher
component is stopped.

The test now passes for me locally.

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

n/a