[BUG]: fix metadata key delete bug in test segment (#4112)
## Description of changes

This fixes a bug in the test segment implementation.

Prior to this change, deletes did not work when metadata only had a
single key. Imagine the case:

1. A record is added with 1 metadata key: `{"a": false}`
2. The record is updated with the metadata `{"a": Null}`

When `merge_meta()` is called during 2., `new_meta` ended up being
`None` because it was empty. This triggered the wrong `match` arm and
ended up returning the old metadata.

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

I checked in a proptest repro.

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

n/a