[BUG]: SysDb should update `TotalRecordsPostCompaction` and `SizeBytesPostCompaction` when flushing with GC enabled (#4131)
## Description of changes

When garbage collection was enabled on the SysDb, it did not properly
update `TotalRecordsPostCompaction` and `SizeBytesPostCompaction` when
`FlushCollectionCompaction` was called. This is now fixed.

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

Updated the mock and test to make sure the values are passed through
correctly.

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

n/a