[ENH]  Shim the rust log client to be able to talk to either log service. (#4124)
## Description of changes

This PR makes the gRPC-backed log client open a client and an alternate
client.  The alternate client will be used iff the alt_host is set and
either use_alt_host_for_everything is set or use_alt_for_collections
contains the collection.

The alternative was to try and fork our configs, have two parallel
stacks, set that up, and tear that down.

## Test plan

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
* Config variables are updated.