[ENH] Add last compaction time seconds to sysdb on flush (#4151)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- For both the version file and no version file route add the last
compaction time seconds to the sysdb column created already.
 - New functionality
   - None

## Test plan
*How are these changes tested?*
Updated the test for the non-version file path to insert and check the
time
The non-version file path does not have a clean way to be tested at this
moment, we confirm the field is returned and audit it manually.
Need to add
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None