[CLN] Rename compact orchestrator field (#4231)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- Rename the `collection_logical_size_bytes` field to
`collection_logical_size_delta_bytes`, and only track the collection
size delta with this field. The aggregated collection size is computed
right before flushing to sysdb. This change aims to improve the clarity
of the code.
 - New functionality
   - ...

## Test plan
*How are these changes tested?*

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

Co-authored-by: Sicheng Pan <sicheng@trychroma.com>