[BUG]: use absolute cutoff time when listing collections to garbage collect (#4245)
## Description of changes

We currently send a relative cutoff time in seconds to the Sysdb when it
expects an absolute cutoff time. This updates the garbage collection
service to send an absolute cutoff time instead.

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

n/a