[CHORE] remove unused page, page_size, and sort args on get (#4204)
## Description of changes

This PR removes the unused page, page_size, and sort arguments on get.
IT also standardizes around the types for Include for get and query
instead of using lists. The async get references returned embeddings as
part of include, which goes against what the docs say here:
https://docs.trychroma.com/docs/querying-collections/query-and-get#choosing-which-data-is-returned
so that is also fixed

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*