[ENH]  Make snapshot be recursive. (#4085)
Before this PR, the status quo was to stop at a depth of 2.  This allows
for a small-ish log.  With this change, the snapshots can be of
arbitrary depth, although in practice they will be log(n)/log(1846)
levels deep.

The motivation for this patch is the realization that 1846**2 gives just
under 4 days of logging before the manifest begins to grow again.