[DOC]: Update improper usage of `$contains` operator in example code and add test to validate it (#4096)
## Description of changes

In various files in the codebase, the string: `{$contains: {"text":
"hello"}}` was used as an example of a `where_document` query. This is
ill-formed for two reasons: (i) `$contains` does not appear in quotation
marks (OK in JS, not in python); (ii) the structure does not conform to
the `WhereDocument` Type: `$contains` cannot be the key for a value that
is a dictionary of two strings. See:
```
WhereDocumentOperator = Union[
    Literal["$contains"], Literal["$not_contains"], LogicalOperator
]
WhereDocument = Dict[WhereDocumentOperator, Union[str, List["WhereDocument"]]]
```
in
[base_types.py](https://github.com/chroma-core/chroma/blob/main/chromadb/base_types.py)

Regrettably, this snippet was referenced in the API docs for
`langchain-chroma`, which is what led me to find and make this fix.

*Summarize the changes made by this PR.*
* Replace occurrences of `{$contains: {"text": "hello"}}` with
`{"$contains": "hello"}` in any comments or documentation.
* Add a failing test for `{"$contains": {"text": "hello"}}`

## Test plan
*How are these changes tested?*

* I added a test (failure) case for the ill-formed `{$contains: {"text":
"hello"}}`. After setting up my basic environment, I was able to run the
unit test, and see that it passed. However, I've been struggling to get
a full development environment setup, so I have not been able to fully
run all the tests.

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

---------

Co-authored-by: Really Him <hesereallyhim@proton.me>