[ENH] add regeneration script for js library (#4115)
## Description of changes
This PR updates the regeneration script for the JS library to fix
failing generation on the new openapi spec. It also overhauls the used
functions, renaming the V2 references back to the normal references as
per the openapi spec.

 - Improvements & Bug fixes
   - ...
 - New functionality
   - ...

## Test plan
*How are these changes tested?*

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

---------

Co-authored-by: Robert Escriva <robert@trychroma.com>
Co-authored-by: Philip I. Thomas <philip@trychroma.com>