[DOC]: Fix docstring defaults for query filters (#4282)
## Description of changes

The docstrings for `where` and `where_document` indicated that they
default to `{}`. This is inconsistent with the actual type definition,
and additionally confusing because sending `None` is OK, but sending
`{}` is not.

So, this is a small find-and-replace for these instances.

## NOTE

This also comes up with respect to the `headers` in the [Docs
site](https://github.com/chroma-core/chroma/blob/069ca5fb5223a1e466244f3c6e858c8e183ce233/docs/docs.trychroma.com/markdoc/content/reference/python/client.md?plain=1#L66)
regarding the HTTP Client. In this case, the docs are clearly
inconsistent with the code right above it, but I wasn't sure if I should
change it, because `None` gets converted to `{}` in sending the request.
I'm still inclined to correct it to match the actual code.

## Test plan
*How are these changes tested?*

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

Co-authored-by: Really Him <hesereallyhim@proton.me>