[CHORE]: Discord - link to channel (#4363)
## Description of changes

This is just closing the loop on #4130 , which seems to have gotten out
of sync with my local branch, so the revision requests I was responding
to weren't getting pushed up, so I'll just close that one. Also, I
noticed that in the meantime @jeffchuber already merged a PR #4286 with
my changes, the only difference being that this adds a link to the
channel. 🤷

So this is a really trivial PR, just making it so that I can close out
the other one and maybe you like this minuscule enhancement.
@jairad26

Co-authored-by: Really Him <hesereallyhim@proton.me>