[DOC]: Fix broken Next/Prev link buttons (#4310)
## Description of changes

Fixes #4215 

On certain pages, specifically those that had a top section with no
subsection, the Next/Prev functionality was sometimes breaking. This was
due to the way the links and the slugs were connected and from the
configuration of the Sidebar and the its components.

## Fixes

It turns out that `path` can be computed from `slug` by just prepending
a `/`, so I removed the `path` property and refactored it to add some
small logic to convert slug to path, simplifying the component
properties.

In addition, I found that some links to the python-thin-client were
broken, due to using relative paths, so I just did a quick fix by using
the absolute path instead. The only downside is that if you are working
in an IDE, the absolute link won't work within the IDE. There are ways
to patch this, or configure editors to find the right path, if people
think this is worth addressing.

## Test plan
*How are these changes tested?*

Run the app locally, validate the cases that are mentioned as breaking
in the Issue referenced above are working now, check other links for
sanity, check links to python-thin-clint, check links to "Edit this page
on GitHub".

Built the app locally and no failures or problems.

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

---------

Co-authored-by: Really Him <hesereallyhim@proton.me>