[ENH] Introduce + consume foyer.obtain() (#4362)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- Removed the write_mutex that block manager had. It does not have any
correctness bearings. It was simply there so that we don't reinsert the
same entry to the cache again. Because we clone and handout the block
(internally a block is an Arc<> so clone is cheap) it should be fine to
overwrite the cache with the same value again.
 - New functionality
- foyer.obtain() deduplicates disk access to the same key so might be
more performant under concurrency. The trade-off is that you have to
clone the key which for us is a 16 byte uuid.

## Test plan
*How are these changes tested?*
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None