[CHORE] Fix flaky default_ef test (#4376)
## Description of changes

https://github.com/chroma-core/chroma/actions/runs/14651962766/job/41119672498

In this run, hypothesis fails this test because too much has been
filtered out. this is because the test fetches all providers, and then
filters to the available ones. a better solution would be to directly
access the available providers

_Summarize the changes made by this PR._

- Improvements & Bug fixes
  - ...
- New functionality
  - ...

## Test plan

_How are these changes tested?_

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes

_Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
section](https://github.com/chroma-core/chroma/tree/main/docs/docs.trychroma.com)?_